restructuring scripts
